Refactor classes to standard package name
