+++ ShopperInsights/src/main/java/com/braintreepayments/api/BuyerPhone.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.shopperinsights
+package com.braintreepayments.api
 
 /**
  * Representation of a user phone number.
