+++ ShopperInsights/src/main/java/com/braintreepayments/api/PaymentReadyApi.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.shopperinsights
+package com.braintreepayments.api
 
 import org.json.JSONObject
