+++ ShopperInsights/src/main/java/com/braintreepayments/api/ShopperInsightsCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.shopperinsights
+package com.braintreepayments.api
 
 /**
  * A callback that returns information on whether someone is a PayPal or a Venmo shopper.
