+++ ShopperInsights/src/main/java/com/braintreepayments/api/ShopperInsightsRequest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.shopperinsights
+package com.braintreepayments.api
 
 /**
  * Data class representing a request for shopper insights.
