+++ ShopperInsights/src/main/java/com/braintreepayments/api/ShopperInsightsResult.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.shopperinsights
+package com.braintreepayments.api
 
 /**
  * The result object returned when insights about a shopper is requested.
